Stac and StacDataOpener added
