+++ test/cassettes/test_store/StacDataStoreTest.test_describe_data.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:01 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:01 GMT
       Source-Age:
-      - '19'
+      - '0'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -57,19 +57,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '1'
+      - '0'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 7f637283b237c2d2b2fb98b7774bf55dae35a1f4
+      - cc4e52f5772183c907b165bbe79dac7e1fbdf56b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320054-CPH
+      - cache-fra-eddf8230084-FRA
       X-Timer:
-      - S1714749928.187199,VS0,VE1
+      - S1716541081.218784,VS0,VE148
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:01 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:01 GMT
       Source-Age:
-      - '19'
+      - '0'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -129,19 +129,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '1'
+      - '0'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - debf3efebab98abdfaadf704dbf43d99a1a559f1
+      - 7d9c4406f2cef79e17bf1575729e7383d9404a7f
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320040-CPH
+      - cache-fra-eddf8230105-FRA
       X-Timer:
-      - S1714749928.216734,VS0,VE1
+      - S1716541081.427758,VS0,VE90
       X-XSS-Protection:
       - 1; mode=block
     status:
