+++ test/cassettes/test_store/StacDataStoreTest.test_do_bboxes_intersect.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:01 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:01 GMT
       Source-Age:
-      - '19'
+      - '0'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - c502af8f8686daf09f8bafaf96bd1c77a6730167
+      - cfd8240f6de70fcc5e27b454b739228b590024f7
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320049-CPH
+      - cache-fra-eddf8230022-FRA
       X-Timer:
-      - S1714749928.005785,VS0,VE1
+      - S1716541082.587935,VS0,VE1
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:01 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:01 GMT
       Source-Age:
-      - '18'
+      - '0'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 4bb2db2a031b7ea0ab81a24a0aa9a7f03a7e3c63
+      - ce9ecafb173bb99a11def85f054f61634f062409
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320026-CPH
+      - cache-fra-eddf8230092-FRA
       X-Timer:
-      - S1714749928.037855,VS0,VE1
+      - S1716541082.639785,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
