+++ test/cassettes/test_store/StacDataStoreTest.test_get_data_ids_optional_args.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:28 GMT
+      - Fri, 24 May 2024 08:58:03 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 10 May 2024 07:14:28 GMT
+      - Fri, 24 May 2024 09:03:03 GMT
       Source-Age:
-      - '5'
+      - '3'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - f31a7b7cf924941255a352cd1457a3efdfb976f6
+      - 43a03215fe02f20bcac461d8e94994249adbbb5c
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A385:105F80C:663DC822
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-fra-eddf8230144-FRA
+      - cache-fra-eddf8230038-FRA
       X-Timer:
-      - S1715324969.917206,VS0,VE1
+      - S1716541084.948158,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:28 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 10 May 2024 07:14:28 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '5'
+      - '2'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 5b8dd2955737a6c17d1d2ca6ed853f70aa005799
+      - b57a476146d36aed2a09964947641803a2dfe1b8
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A39E:105F82D:663DC823
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-fra-eddf8230031-FRA
+      - cache-fra-eddf8230109-FRA
       X-Timer:
-      - S1715324969.988517,VS0,VE1
+      - S1716541084.000581,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -203,13 +203,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:29 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"ddd340bc27c120dd2e43868bcde0510a326a6223dac1b0c47c05100e20d1397e"'
       Expires:
-      - Fri, 10 May 2024 07:14:29 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '5'
+      - '3'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -223,15 +223,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 7518c1590b56ad5b652afcb45cd5eeaba2598786
+      - 80f4277c664be9f04a70abb4f53faf2a8bf8a49b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 27EE:3635F7:5B574A:611F18:663DC823
+      - 7F34:D7055:26F806E:29015C2:66505262
       X-Served-By:
-      - cache-fra-eddf8230145-FRA
+      - cache-fra-eddf8230068-FRA
       X-Timer:
-      - S1715324969.068276,VS0,VE1
+      - S1716541085.542760,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -287,13 +287,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:29 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"3263faca1f19517d02862736694703cc8519bed9344039cace8aa2c5f9379bcf"'
       Expires:
-      - Fri, 10 May 2024 07:14:29 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '5'
+      - '2'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -307,15 +307,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 9c331e7565f05fdb7ea8ac388b68aa75a8409dc9
+      - 596b40fde50561635b4eabb70cfcac5c7ca50f1b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - E09A:28EC35:F3998B:101ED8D:663DC822
+      - 5554:366126:1E0E130:1F94C3C:66505262
       X-Served-By:
-      - cache-fra-eddf8230121-FRA
+      - cache-fra-eddf8230085-FRA
       X-Timer:
-      - S1715324969.163934,VS0,VE1
+      - S1716541085.595749,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -387,13 +387,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:29 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"80ec96bc0acf2e604a03f109bd730426aa82e442d44946231cbe82a531b944f7"'
       Expires:
-      - Fri, 10 May 2024 07:14:29 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '5'
+      - '2'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -407,15 +407,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - e97ab3512577b2eb114cc3c89fff6e80ac949bc8
+      - 924b5c0df9cc8013ed262275a00bfa76f85e4466
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 3D42:208DB8:E6CE03:F4739A:663DC824
+      - 6B90:37C204:1CC468F:1E3D692:66505261
       X-Served-By:
-      - cache-fra-eddf8230113-FRA
+      - cache-fra-eddf8230136-FRA
       X-Timer:
-      - S1715324969.238517,VS0,VE1
+      - S1716541085.648799,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -487,13 +487,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:29 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"726870312c74ead0b10c3125045c301e8600929684c49447d64c2db72dc779fc"'
       Expires:
-      - Fri, 10 May 2024 07:14:29 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '5'
+      - '2'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -507,15 +507,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 7f11a08069ece623357e7a8f97d1e6c9d03c5c77
+      - 613874fe1deee9cbf96501e98ee47f5d774aa9ad
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 25B8:28EC35:F399B5:101EDC3:663DC823
+      - D9D6:198FA3:2BF2E8A:2E48E77:66505263
       X-Served-By:
-      - cache-fra-eddf8230096-FRA
+      - cache-fra-eddf8230116-FRA
       X-Timer:
-      - S1715324969.313457,VS0,VE1
+      - S1716541085.701349,VS0,VE5
       X-XSS-Protection:
       - 1; mode=block
     status:
