+++ test/cassettes/test_store/StacDataStoreTest.test_get_data_ids_optional_args_empty_args.yaml
@@ -41,11 +41,11 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
       - '3'
       Strict-Transport-Security:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - df710316c85a5940f86fd1e2cacb1bcca6fbdc5c
+      - 42b96759653078ba3d851b75d1569d5991323b9d
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A385:105F80C:663DC822
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-fra-eddf8230084-FRA
+      - cache-fra-eddf8230157-FRA
       X-Timer:
-      - S1715324966.145067,VS0,VE1
+      - S1716541085.772798,VS0,VE13
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '2'
+      - '3'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 826bab83ce24f34d5b8d70446967659610b61efd
+      - f28012f1a58bb56c04a1367713e80a87c479724b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A39E:105F82D:663DC823
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-fra-eddf8230061-FRA
+      - cache-fra-eddf8230043-FRA
       X-Timer:
-      - S1715324966.219325,VS0,VE1
+      - S1716541085.837286,VS0,VE20
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -203,13 +203,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"ddd340bc27c120dd2e43868bcde0510a326a6223dac1b0c47c05100e20d1397e"'
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '2'
+      - '3'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -223,15 +223,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 8afd1c60fb3b418ac9ba0e386a1a09fef0b27d0a
+      - 43137188ffd96628e284b279b92cfb35254a23e6
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 27EE:3635F7:5B574A:611F18:663DC823
+      - 7F34:D7055:26F806E:29015C2:66505262
       X-Served-By:
-      - cache-fra-eddf8230073-FRA
+      - cache-fra-eddf8230112-FRA
       X-Timer:
-      - S1715324966.292536,VS0,VE1
+      - S1716541085.908000,VS0,VE3
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -287,13 +287,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:04 GMT
       ETag:
       - '"3263faca1f19517d02862736694703cc8519bed9344039cace8aa2c5f9379bcf"'
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:04 GMT
       Source-Age:
-      - '2'
+      - '3'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -307,15 +307,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 76c9632d0d0f8d5d860680b358cd120d2630b358
+      - 7edad6118331482f87cb5631c14011f83a5ec9d3
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - E09A:28EC35:F3998B:101ED8D:663DC822
+      - 5554:366126:1E0E130:1F94C3C:66505262
       X-Served-By:
-      - cache-fra-eddf8230021-FRA
+      - cache-fra-eddf8230119-FRA
       X-Timer:
-      - S1715324966.366755,VS0,VE1
+      - S1716541085.968645,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -387,11 +387,11 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:05 GMT
       ETag:
       - '"80ec96bc0acf2e604a03f109bd730426aa82e442d44946231cbe82a531b944f7"'
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:05 GMT
       Source-Age:
       - '2'
       Strict-Transport-Security:
@@ -407,15 +407,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - c8f2191e668042a0e0ba591bd521991c89567264
+      - 0902ce67ae5000a74b704aa27b5ffe22495c1c96
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 3D42:208DB8:E6CE03:F4739A:663DC824
+      - 6B90:37C204:1CC468F:1E3D692:66505261
       X-Served-By:
-      - cache-fra-eddf8230136-FRA
+      - cache-fra-eddf8230075-FRA
       X-Timer:
-      - S1715324966.438652,VS0,VE1
+      - S1716541085.021618,VS0,VE3
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -487,11 +487,11 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:09:26 GMT
+      - Fri, 24 May 2024 08:58:05 GMT
       ETag:
       - '"726870312c74ead0b10c3125045c301e8600929684c49447d64c2db72dc779fc"'
       Expires:
-      - Fri, 10 May 2024 07:14:26 GMT
+      - Fri, 24 May 2024 09:03:05 GMT
       Source-Age:
       - '2'
       Strict-Transport-Security:
@@ -507,15 +507,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 41ef574beb513786b84774c7440a2b5bbc8f1e2e
+      - 16c67c6c8ff8caa90c9bf6a527d741e2b9c454b8
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 25B8:28EC35:F399B5:101EDC3:663DC823
+      - D9D6:198FA3:2BF2E8A:2E48E77:66505263
       X-Served-By:
-      - cache-fra-eddf8230103-FRA
+      - cache-fra-eddf8230101-FRA
       X-Timer:
-      - S1715324967.513187,VS0,VE1
+      - S1716541085.078486,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
