+++ test/cassettes/test_store/StacDataStoreTest.test_get_data_opener_ids.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:05 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:05 GMT
       Source-Age:
-      - '19'
+      - '4'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 61f6709e4d3a89ae6c839ffd09478569b65521ee
+      - dd1315af903842a5a4f2aa7d97161830c5faf6ee
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320044-CPH
+      - cache-fra-eddf8230053-FRA
       X-Timer:
-      - S1714749928.374867,VS0,VE1
+      - S1716541085.146954,VS0,VE1
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:28 GMT
+      - Fri, 24 May 2024 08:58:05 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:28 GMT
+      - Fri, 24 May 2024 09:03:05 GMT
       Source-Age:
-      - '19'
+      - '4'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 76b94382cbeb9fd3aec4533d48822e962c1c9042
+      - b8bef633200b02546bfcc68a451f3f7ed8dd2696
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320034-CPH
+      - cache-fra-eddf8230114-FRA
       X-Timer:
-      - S1714749928.413448,VS0,VE1
+      - S1716541085.231558,VS0,VE14
       X-XSS-Protection:
       - 1; mode=block
     status:
