+++ test/cassettes/test_store/StacDataStoreTest.test_get_item_collection_open_params.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:19 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 10 May 2024 07:31:19 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '7'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -57,19 +57,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - a9d60e6c9d959d30d790bf2c6118ee60c9fed882
+      - 55c7d0e1f5b540bdabc8c174bba555c45059976d
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A385:105F80C:663DC822
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-fra-eddf8230132-FRA
+      - cache-fra-eddf8230088-FRA
       X-Timer:
-      - S1715325980.835170,VS0,VE160
+      - S1716541088.230992,VS0,VE3
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:20 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 10 May 2024 07:31:20 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '7'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -129,19 +129,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - add68d93fc4ee1e5942e9ab0447af78fe2e2df21
+      - c27a6bb782adbd2c5297c7e42e8e9c61025ef9a9
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 5BAC:6DBF6:F7A39E:105F82D:663DC823
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-fra-eddf8230046-FRA
+      - cache-fra-eddf8230113-FRA
       X-Timer:
-      - S1715325980.069099,VS0,VE96
+      - S1716541088.285931,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -203,13 +203,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:20 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - '"ddd340bc27c120dd2e43868bcde0510a326a6223dac1b0c47c05100e20d1397e"'
       Expires:
-      - Fri, 10 May 2024 07:31:20 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '6'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -219,19 +219,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - f5742da6134986647fb7cfea551f6a2dfb49344e
+      - ec613abd59875f32c6f6d47d50a38a65209d8806
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 27EE:3635F7:5B574A:611F18:663DC823
+      - 7F34:D7055:26F806E:29015C2:66505262
       X-Served-By:
-      - cache-fra-eddf8230049-FRA
+      - cache-fra-eddf8230031-FRA
       X-Timer:
-      - S1715325980.257083,VS0,VE142
+      - S1716541088.341966,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -287,13 +287,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:20 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - '"3263faca1f19517d02862736694703cc8519bed9344039cace8aa2c5f9379bcf"'
       Expires:
-      - Fri, 10 May 2024 07:31:20 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '6'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -303,19 +303,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 079c3cbe8827c5ff06f44ec402ec96739a61bfe7
+      - 0bee4b76a9688fad9a7677deefc38c8fa4445a2b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - E09A:28EC35:F3998B:101ED8D:663DC822
+      - 5554:366126:1E0E130:1F94C3C:66505262
       X-Served-By:
-      - cache-fra-eddf8230158-FRA
+      - cache-fra-eddf8230113-FRA
       X-Timer:
-      - S1715325980.470818,VS0,VE143
+      - S1716541088.396056,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -387,13 +387,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:20 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - '"80ec96bc0acf2e604a03f109bd730426aa82e442d44946231cbe82a531b944f7"'
       Expires:
-      - Fri, 10 May 2024 07:31:20 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '6'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -403,19 +403,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - b05526f9ee3ca1158cb6f0b081c492ba1c21546e
+      - 61af7c3c3137fb3fab8a126f9a287992798de493
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 3D42:208DB8:E6CE03:F4739A:663DC824
+      - 6B90:37C204:1CC468F:1E3D692:66505261
       X-Served-By:
-      - cache-fra-eddf8230153-FRA
+      - cache-fra-eddf8230038-FRA
       X-Timer:
-      - S1715325981.686865,VS0,VE163
+      - S1716541088.448608,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -487,13 +487,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 10 May 2024 07:26:21 GMT
+      - Fri, 24 May 2024 08:58:08 GMT
       ETag:
       - '"726870312c74ead0b10c3125045c301e8600929684c49447d64c2db72dc779fc"'
       Expires:
-      - Fri, 10 May 2024 07:31:21 GMT
+      - Fri, 24 May 2024 09:03:08 GMT
       Source-Age:
-      - '0'
+      - '6'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -503,19 +503,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - bc47486156ae4857de917f82ce094980c4c96a09
+      - 54a1c8fefd382bdd7d4c58a2d67cd9416e6e4951
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 25B8:28EC35:F399B5:101EDC3:663DC823
+      - D9D6:198FA3:2BF2E8A:2E48E77:66505263
       X-Served-By:
-      - cache-fra-eddf8230110-FRA
+      - cache-fra-eddf8230104-FRA
       X-Timer:
-      - S1715325981.927171,VS0,VE132
+      - S1716541089.502936,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
