+++ test/cassettes/test_store/StacDataStoreTest.test_get_open_data_params_schema.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:25 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:25 GMT
       Source-Age:
-      - '20'
+      - '24'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 441d80997bf39fb2f2cc7425e54a9130a245059a
+      - 4d863b1f05ad5121c29f279287d00045238e29fd
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320035-CPH
+      - cache-fra-eddf8230063-FRA
       X-Timer:
-      - S1714749929.060379,VS0,VE1
+      - S1716541105.227431,VS0,VE6
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:25 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:25 GMT
       Source-Age:
-      - '19'
+      - '24'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - b7b4c7d78fe782beb00c1b4e22c4a0a99d3c8814
+      - ed2f870f753d07fb2b2620ae22ce730d36d68e66
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320045-CPH
+      - cache-fra-eddf8230050-FRA
       X-Timer:
-      - S1714749929.090788,VS0,VE1
+      - S1716541105.284938,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
