+++ test/cassettes/test_store/StacDataStoreTest.test_get_search_params_schema.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:25 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:25 GMT
       Source-Age:
-      - '20'
+      - '24'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -61,15 +61,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 51bd0e28674149f8a6511e7220709c5d707cd6b8
+      - e351a007ef560d2bbc1fe4f47694b16ce024682b
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320046-CPH
+      - cache-fra-eddf8230025-FRA
       X-Timer:
-      - S1714749929.174729,VS0,VE1
+      - S1716541105.392760,VS0,VE10
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:25 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:25 GMT
       Source-Age:
-      - '20'
+      - '24'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 674f86cc5615b1a123279ed51a7cc33bc98cd2dc
+      - dc2d8117fcae93d0eee856b23414db59fb8cf099
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320054-CPH
+      - cache-fra-eddf8230022-FRA
       X-Timer:
-      - S1714749929.221389,VS0,VE1
+      - S1716541105.455495,VS0,VE3
       X-XSS-Protection:
       - 1; mode=block
     status:
