+++ test/cassettes/test_store/StacDataStoreTest.test_is_datetime_in_range.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Tue, 14 May 2024 09:08:52 GMT
+      - Fri, 24 May 2024 08:58:27 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Tue, 14 May 2024 09:13:52 GMT
+      - Fri, 24 May 2024 09:03:27 GMT
       Source-Age:
-      - '0'
+      - '26'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -57,19 +57,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 676dd71a8ff122cdec17e94a3ab8a088bd54d819
+      - 30c5bb778ba522a9623085110005f936fae503e5
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - D0EE:148AA2:18A7844:1A1CD48:66432132
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-fra-eddf8230108-FRA
+      - cache-fra-eddf8230026-FRA
       X-Timer:
-      - S1715677732.487448,VS0,VE143
+      - S1716541108.575831,VS0,VE5
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Tue, 14 May 2024 09:08:52 GMT
+      - Fri, 24 May 2024 08:58:27 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Tue, 14 May 2024 09:13:52 GMT
+      - Fri, 24 May 2024 09:03:27 GMT
       Source-Age:
-      - '0'
+      - '26'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -127,21 +127,21 @@ interactions:
       Via:
       - 1.1 varnish
       X-Cache:
-      - MISS
+      - HIT
       X-Cache-Hits:
-      - '0'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 36989abd8874aadf2e50450fc56334d4fe7c358e
+      - 4d64c2f2eaea649f05761540d6beede2d2684862
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - 6A22:312E01:32FCAB4:35F3692:66432A24
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-fra-eddf8230147-FRA
+      - cache-fra-eddf8230033-FRA
       X-Timer:
-      - S1715677733.682246,VS0,VE177
+      - S1716541108.632084,VS0,VE3
       X-XSS-Protection:
       - 1; mode=block
     status:
