+++ test/cassettes/test_store/StacDataStoreTest.test_open_data.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:27 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:27 GMT
       Source-Age:
-      - '20'
+      - '26'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -57,19 +57,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '2'
+      - '1'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - abe4f0d57fc2868cde252d36a04a95c3849e3e1e
+      - f34e0980e36c76b1fc21a0ecd90f45e3358561b3
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320041-CPH
+      - cache-fra-eddf8230133-FRA
       X-Timer:
-      - S1714749929.388680,VS0,VE0
+      - S1716541108.696285,VS0,VE1
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:27 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:27 GMT
       Source-Age:
-      - '20'
+      - '26'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 5c412d7c8ec5bd27cbb15106cd2700a64e41477b
+      - b4d50e30196babc58432ffbc42360c1eeff3b209
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320044-CPH
+      - cache-fra-eddf8230046-FRA
       X-Timer:
-      - S1714749929.425456,VS0,VE1
+      - S1716541108.753686,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
