+++ test/cassettes/test_store/StacDataStoreTest.test_search_data.yaml
@@ -41,13 +41,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:27 GMT
       ETag:
       - W/"acb7a8d6636e24e32f4018c14f1c4ff418a82567b2746560f9eae6ad97a48a54"
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:27 GMT
       Source-Age:
-      - '20'
+      - '27'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -57,19 +57,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '1'
+      - '2'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 02e8e14da5f5ffad99ec2622434aceeec44f85f3
+      - b2e60af45391757ac43a97192ecca02b608550b0
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F67:1E6FB08:6634FE90
+      - 36D3:107820:22CCEEE:24A2ADC:665050BD
       X-Served-By:
-      - cache-cph2320051-CPH
+      - cache-fra-eddf8230147-FRA
       X-Timer:
-      - S1714749930.603023,VS0,VE1
+      - S1716541108.962343,VS0,VE0
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -113,13 +113,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Fri, 03 May 2024 15:25:29 GMT
+      - Fri, 24 May 2024 08:58:28 GMT
       ETag:
       - '"e74ebcbc46d43c5b693ecb995381fbeba03583627e6d65b21ed7678a10d94729"'
       Expires:
-      - Fri, 03 May 2024 15:30:29 GMT
+      - Fri, 24 May 2024 09:03:28 GMT
       Source-Age:
-      - '20'
+      - '26'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -133,15 +133,15 @@ interactions:
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 0ee4551d52e42dd8d6e2086c23c5481170380f35
+      - fd273ea425c62f9e42d160706e4119b50b6a3c53
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
-      - BACE:2E8F4B:1CE9F92:1E6FB29:6634FE90
+      - 36CD:107820:22CCF1B:24A2B05:665050BD
       X-Served-By:
-      - cache-cph2320055-CPH
+      - cache-fra-eddf8230087-FRA
       X-Timer:
-      - S1714749930.647178,VS0,VE1
+      - S1716541108.014349,VS0,VE2
       X-XSS-Protection:
       - 1; mode=block
     status:
